#9 Added support for two cameras
Two cameras can now be used simuntaniosly. The plan is to make an implementation that can use n cameras in the future. The problem is the dynamic allocation of the ports and the dynamic allocation of the gui elements for the video streaming. But more on that perhaps in the future.