#12 Started implementing email notifications
EmailSender class was created which sends the image to the given receiver. The class constructor receives the camera number and data from the camera in byte format, saves the image to temp folder and sends an email with the image as attachment.
Until a better solution is found, the email is sent from newly created gmail account and through gmail's SMTP server.